libzbd: Fix 32-bit overflow when using 4GB zone size
Even though zone_size is 64-bit, zone_sectors is only 32-bit
(since that is what the BLKGETZONESZ ioctl wants).
Promote zone_sectors before doing the bit shift, in order to
avoid overflow when using large zone sizes.

Reported-by: Subrahmanya.Lingappa <subrahmanya.lingappa@wdc.com>

+++ lib/zbd.c
@@ -256,7 +256,7 @@ static struct zbd_info *zbd_do_get_info(int fd, char *devname)
 		goto err;
 	}
 	zbdi->zone_sectors = zone_sectors;
-	zbdi->zone_size = zone_sectors << 9;
+	zbdi->zone_size = (size_t)zone_sectors << 9;
 
 	/* Get number of zones */
 	ret = ioctl(fd, BLKGETNRZONES, &nr_zones);
